Switch away from getters/setters
Getters and setters are, well, problematic in Rust. Initially I implemented
them to be able to validate date before it is accepted as change, but I
don't think it's actually a good way to do that. Instead I might consider
adding validate function later that one can use to check if all properties
are valid.